backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-dotenv==1.0.0
google-api-python-client==2.86.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
numpy==1.24.3
pandas==1.5.3
gspread==5.12.1
oauth2client==4.1.3
requests==2.31.0
